Implementadas las tarjetas, método que sume lo acumulado pagado mediante tarjetas. Sólo se para el programa desde el modo admin.

+++ src/main/java/daw/poo/maquinavending/gabrielallendejuanjosebazan/Productos.java
@@ -37,7 +37,7 @@ public class Productos {
         this.nombreProducto = nombreProducto;
     }
 
-    public double getPrecioProducto() {
+    public int getPrecioProducto() {
         return precioProducto;
     }
